// types.d.ts
export interface HNItem {
    id: number;
    deleted?: boolean;
    type?: 'job' | 'story' | 'comment' | 'poll' | 'pollopt';
    by?: string;
    time?: number;
    text?: string;
    dead?: boolean;
    parent?: number;
    poll?: number;
    kids?: number[];
    url?: string;
    score?: number;
    title?: string;
    parts?: number[];
    descendants?: number;
}
export interface HNUser {
    id: string;
    delay?: number;
    created: number;
    karma: number;
    about?: string;
    submitted?: number[];
}
export interface StoryDisplay {
    id: number;
    title: string;
    url?: string;
    score?: number;
    by?: string;
    time?: number;
    descendants?: number;
    commentsUrl: string;
}
export type StoryType = 'top' | 'new' | 'best' | 'ask' | 'show' | 'job';
export interface AppConfig {
    limit: number;
    storyType: StoryType;
    openInBrowser: boolean;
}
//# sourceMappingURL=types.d.ts.map